fix: modify profile
