+++ index.html
@@ -1,10 +1,13 @@
-<!DOCTYPE html>
+<!doctype html>
 <html lang="zh-TW">
   <head>
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <title>STYLiSH</title>
     <link rel="preconnect" href="https://fonts.googleapis.com" />
+
+    <link href="./src/css/profile.css" rel="stylesheet" />
+
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
     <link
       href="https://fonts.googleapis.com/css2?family=Noto+Sans+TC:wght@400;700&display=swap"
