+++ src/css/profile.css
@@ -1,0 +1,72 @@
+.profile {
+  padding: 0 20px 200px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  text-align: center;
+}
+
+.profile__title {
+  padding-bottom: 16px;
+  border-bottom: 1px solid #979797;
+  font-size: 24px;
+  font-weight: bold;
+}
+
+.profile__content {
+  margin-top: 24px;
+}
+
+.profile__info {
+  font-size: 18px;
+  font-weight: bold;
+  margin-bottom: 15px;
+}
+
+.profile__form {
+  width: 100%;
+  max-width: 450px;
+  display: flex;
+  flex-direction: column;
+}
+
+.profile__label {
+  font-size: 18px;
+  color: #333;
+  font-weight: 200;
+  margin-bottom: 5px;
+  align-self: flex-start;
+}
+
+.profile__hint {
+  color: #ccc;
+  text-decoration: underline;
+  cursor: pointer;
+}
+
+.profile__input {
+  width: 100%;
+  height: 40px;
+  font-size: 18px;
+  margin-bottom: 20px;
+  border-radius: 16px;
+  padding: 0px 10px;
+}
+
+.profile__button {
+  width: 100%;
+  height: 60px;
+  margin-top: 29px;
+  border: solid 1px #979797;
+  background-color: black;
+  font-size: 20px;
+  letter-spacing: 4px;
+  color: white;
+  cursor: pointer;
+}
+
+@media screen and (max-width: 1279px) {
+  .profile {
+    padding: 0px 20px 240px;
+  }
+}
